+++ biom/err.py
@@ -75,7 +75,8 @@ OBSMDSIZE = "Size of observation metadata differs from matrix size!"
 SAMPMDSIZE = "Size of sample metadata differs from matrix size!"
 
 
-def _test_empty(t):
+# _zz_ so the sort order places this test last
+def _zz_test_empty(t):
     """Check if t is empty"""
     return t.is_empty()
 
@@ -250,8 +251,9 @@ class ErrorProfile(object):
         if not args:
             args = self._test.keys()
 
-        for errtype in args:
+        for errtype in sorted(args):
             test = self._test.get(errtype, lambda: None)
+
             if test(item):
                 return self._handle_error(errtype, item)
 
@@ -318,7 +320,7 @@ class ErrorProfile(object):
 
 
 __errprof = ErrorProfile()
-__errprof.register('empty', EMPTY, 'ignore', _test_empty,
+__errprof.register('empty', EMPTY, 'ignore', _zz_test_empty,
                    exception=TableException)
 __errprof.register('obssize', OBSSIZE, 'raise', _test_obssize,
                    exception=TableException)