+++ biom/tests/test_err.py
@@ -11,12 +11,14 @@
 from unittest import TestCase, main
 from copy import deepcopy
 
+import numpy as np
+
 from biom import example_table, Table
 from biom.exception import TableException
-from biom.err import (_test_empty, _test_obssize, _test_sampsize, _test_obsdup,
-                      _test_sampdup, _test_obsmdsize, _test_sampmdsize,
-                      errstate, geterr, seterr, geterrcall, seterrcall,
-                      errcheck, __errprof)
+from biom.err import (_zz_test_empty, _test_obssize, _test_sampsize,
+                      _test_obsdup, _test_sampdup, _test_obsmdsize,
+                      _test_sampmdsize, errstate, geterr, seterr, geterrcall,
+                      seterrcall, errcheck, __errprof)
 
 
 runtime_ep = __errprof
@@ -30,8 +32,8 @@ class ErrModeTests(TestCase):
         self.ex_table = example_table.copy()
 
     def test_test_empty(self):
-        self.assertTrue(_test_empty(Table([], [], [])))
-        self.assertFalse(_test_empty(self.ex_table))
+        self.assertTrue(_zz_test_empty(Table([], [], [])))
+        self.assertFalse(_zz_test_empty(self.ex_table))
 
     def test_test_obssize(self):
         self.assertFalse(_test_obssize(self.ex_table))
@@ -86,6 +88,17 @@ class ErrorProfileTests(TestCase):
         self.ep.test(self.ex_table, 'empty')
         self.assertTrue(isinstance(self.ep.test(self.ex_table, 'obssize'),
                                    TableException))
+
+    def test_test_evaluation_order(self):
+        # issue 813
+        tab = Table(np.array([[1, 2], [3, 4]]), ['A', 'B'], ['C', 'D'])
+        tab._observation_ids = np.array(['A', 'A'], dtype='object')
+        tab._sample_ids = np.array(['B', 'B'], dtype='object')
+
+        self.assertEqual(self.ep.test(tab, 'obsdup', 'sampdup').args[0],
+                         'Duplicate observation IDs')
+        self.assertEqual(self.ep.test(tab, 'sampdup', 'obsdup').args[0],
+                         'Duplicate observation IDs')
 
     def test_state(self):
         self.ep.state = {'all': 'ignore'}